Make AI take rotation into account
When checking for imminent victory, if the board is about to rotate, the
AI makes a copy of the board, adds a piece, and rotates it, letting the
pieces fall, before it checks for victory. This means it detects and
accounts for 4-in-a-rows caused by the rotation of the board. Rotating
is a non-reversible process, so a new board is created each time.